feat(lint): add `msg-value-loop`
