+++ crates/lint/src/sol/low/mod.rs
@@ -9,6 +9,9 @@ use calls_loop::CALLS_LOOP;
 mod delegatecall_loop;
 use delegatecall_loop::DELEGATECALL_LOOP;
 
+mod msg_value_loop;
+use msg_value_loop::MSG_VALUE_LOOP;
+
 mod missing_zero_check;
 use missing_zero_check::MISSING_ZERO_CHECK;
 
@@ -21,6 +24,8 @@ use missing_events_arithmetic::MISSING_EVENTS_ARITHMETIC;
 mod return_bomb;
 use return_bomb::RETURN_BOMB;
 
+mod payable_loop;
+
 mod reentrancy_events;
 use reentrancy_events::REENTRANCY_EVENTS;
 
@@ -28,6 +33,7 @@ register_lints!(
     (BlockTimestamp, early, (BLOCK_TIMESTAMP)),
     (CallsLoop, late, (CALLS_LOOP)),
     (DelegatecallLoop, late, (DELEGATECALL_LOOP)),
+    (MsgValueLoop, late, (MSG_VALUE_LOOP)),
     (MissingEventsAccessControl, late, (MISSING_EVENTS_ACCESS_CONTROL)),
     (MissingEventsArithmetic, late, (MISSING_EVENTS_ARITHMETIC)),
     (MissingZeroCheck, late, (MISSING_ZERO_CHECK)),
