+++ crates/lint/testdata/MsgValueLoop.stderr
@@ -1,0 +1,112 @@
+warning[msg-value-loop]: payable functions should not use `msg.value` inside a loop
+   ╭▸ ROOT/testdata/MsgValueLoop.sol:LL:CC
+   │
+LL │         return msg.value;
+   │                ━━━━━━━━━
+   │
+   ╰ help: https://getfoundry.sh/forge/linting/msg-value-loop
+
+warning[msg-value-loop]: payable functions should not use `msg.value` inside a loop
+   ╭▸ ROOT/testdata/MsgValueLoop.sol:LL:CC
+   │
+LL │             total += msg.value;
+   │                      ━━━━━━━━━
+   │
+   ╰ help: https://getfoundry.sh/forge/linting/msg-value-loop
+
+warning[msg-value-loop]: payable functions should not use `msg.value` inside a loop
+   ╭▸ ROOT/testdata/MsgValueLoop.sol:LL:CC
+   │
+LL │             total += msg.value;
+   │                      ━━━━━━━━━
+   │
+   ╰ help: https://getfoundry.sh/forge/linting/msg-value-loop
+
+warning[msg-value-loop]: payable functions should not use `msg.value` inside a loop
+   ╭▸ ROOT/testdata/MsgValueLoop.sol:LL:CC
+   │
+LL │             total += msg.value;
+   │                      ━━━━━━━━━
+   │
+   ╰ help: https://getfoundry.sh/forge/linting/msg-value-loop
+
+warning[msg-value-loop]: payable functions should not use `msg.value` inside a loop
+   ╭▸ ROOT/testdata/MsgValueLoop.sol:LL:CC
+   │
+LL │             total += msg.value;
+   │                      ━━━━━━━━━
+   │
+   ╰ help: https://getfoundry.sh/forge/linting/msg-value-loop
+
+warning[msg-value-loop]: payable functions should not use `msg.value` inside a loop
+   ╭▸ ROOT/testdata/MsgValueLoop.sol:LL:CC
+   │
+LL │             total += msg.value;
+   │                      ━━━━━━━━━
+   │
+   ╰ help: https://getfoundry.sh/forge/linting/msg-value-loop
+
+warning[msg-value-loop]: payable functions should not use `msg.value` inside a loop
+   ╭▸ ROOT/testdata/MsgValueLoop.sol:LL:CC
+   │
+LL │         for (uint256 i; i < iterations; value = msg.value + i++) {}
+   │                                                 ━━━━━━━━━
+   │
+   ╰ help: https://getfoundry.sh/forge/linting/msg-value-loop
+
+warning[msg-value-loop]: payable functions should not use `msg.value` inside a loop
+   ╭▸ ROOT/testdata/MsgValueLoop.sol:LL:CC
+   │
+LL │         total += msg.value;
+   │                  ━━━━━━━━━
+   │
+   ╰ help: https://getfoundry.sh/forge/linting/msg-value-loop
+
+warning[msg-value-loop]: payable functions should not use `msg.value` inside a loop
+   ╭▸ ROOT/testdata/MsgValueLoop.sol:LL:CC
+   │
+LL │         return msg.value;
+   │                ━━━━━━━━━
+   │
+   ╰ help: https://getfoundry.sh/forge/linting/msg-value-loop
+
+warning[msg-value-loop]: payable functions should not use `msg.value` inside a loop
+   ╭▸ ROOT/testdata/MsgValueLoop.sol:LL:CC
+   │
+LL │             total += msg.value;
+   │                      ━━━━━━━━━
+   │
+   ╰ help: https://getfoundry.sh/forge/linting/msg-value-loop
+
+warning[msg-value-loop]: payable functions should not use `msg.value` inside a loop
+   ╭▸ ROOT/testdata/MsgValueLoop.sol:LL:CC
+   │
+LL │         return msg.value;
+   │                ━━━━━━━━━
+   │
+   ╰ help: https://getfoundry.sh/forge/linting/msg-value-loop
+
+warning[msg-value-loop]: payable functions should not use `msg.value` inside a loop
+   ╭▸ ROOT/testdata/MsgValueLoop.sol:LL:CC
+   │
+LL │         return msg.value;
+   │                ━━━━━━━━━
+   │
+   ╰ help: https://getfoundry.sh/forge/linting/msg-value-loop
+
+warning[msg-value-loop]: payable functions should not use `msg.value` inside a loop
+   ╭▸ ROOT/testdata/MsgValueLoop.sol:LL:CC
+   │
+LL │         return msg.value;
+   │                ━━━━━━━━━
+   │
+   ╰ help: https://getfoundry.sh/forge/linting/msg-value-loop
+
+warning[msg-value-loop]: payable functions should not use `msg.value` inside a loop
+   ╭▸ ROOT/testdata/MsgValueLoop.sol:LL:CC
+   │
+LL │         return msg.value;
+   │                ━━━━━━━━━
+   │
+   ╰ help: https://getfoundry.sh/forge/linting/msg-value-loop
+
